dom/html/test/dialog/test_bug1648877_dialog_fullscreen_denied.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1648877
-->
<head>
  <title>Test for Bug 1648877</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=1648877">Requesting
  fullscreen a dialog element should be denied</a>
<p id="display"></p>
<dialog>
</dialog>
<div style="width: 30px; height:30px" </div>
    
<pre id="test">
<script type="application/javascript">
SimpleTest.waitForExplicitFinish();

function runTest() {
  document.addEventListener("fullscreenchange", () => {
    ok(false, "Should never receive " +
      "a fullscreenchange event in the main window.");
  });

  document.addEventListener('fullscreenerror', (event) => {
    ok(!document.fullscreenElement,
       "Should not grant request if the element is dialog");
    SimpleTest.finish();
  });

  const div = document.querySelector("div");

  div.addEventListener("click", function() {
    const dialog = document.querySelector("dialog");
    dialog.requestFullscreen();
  });

  synthesizeMouseAtCenter(div, {});
}

SimpleTest.waitForFocus(runTest);

</script>
</pre>
</body>
</html>
